feat(integration-evaluation-validation): add BC pretrain config and docs
- bc_pretrain.yaml exposes the warm-start pipeline as tunable config
- README documents the pretrain → train warm-start workflow end to end

Closes #86
Closes #87
Closes #88
Closes #89
Closes #90

+++ training/bc_data.py
@@ -107,9 +107,9 @@ def _load_batches(path: Path, batch_size: int) -> Iterator[Batch]:
     for s in range(0, n, batch_size):
         e = s + batch_size
         yield Batch(
-            obs=torch.from_numpy(obs[s:e]),
+            obs=torch.from_numpy(obs[s:e]).float(),
             targets={h: torch.from_numpy(tgts[h][s:e]).long() for h in _HEAD_NAMES},
-            mc_return=torch.from_numpy(mc_ret[s:e]),
+            mc_return=torch.from_numpy(mc_ret[s:e]).float(),
         )
